testing: moving the helper functions into the internal package
